Change battery sensor to register to callback since library now supports multiple callbacks Move DeconzEvent to hub Bump to v9
