+++ requirements_all.txt
@@ -662,7 +662,7 @@ pycsspeechtts==1.0.2
 # pycups==1.9.73
 
 # homeassistant.components.deconz
-pydeconz==8
+pydeconz==9
 
 # homeassistant.components.zwave
 pydispatcher==2.0.5
